Finished Team Activity up to Step 6
